+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added StringToUpperConverter to convert the given value as string to upper.
 - Added StringCheckToBooleanConverter to execute a check on a single string or list of strings and returns a boolean representation of that result.
 - Added StringCheckToVisibilityConverter to execute a check on a single string or list of strings and returns a Visibility representation of that result.
+- Added NumberCheckToVisibilityConverter to execute a check on a single number or list of numbers and returns a Visibility representation of that result.
 ### Changed
 - Move the parameter for the EqualsToBooleanConverter to a converter property.
 - Change the configurable return value for the EqualsToBooleanConverter from a single to multiple properties.
@@ -54,6 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove NullToVisibilityConverter, thats replaced by IsNullToVisibilityConverter.
 - Remove CountToBooleanConverter, thats replaced by NumberCheckToBooleanConverter.
 - Remove MultiValueToThicknessConverter, thats integrated into the DoubleToThicknessConverter.
+- Remove CountToVisibilityConverter, thats replaced by NumberCheckToVisibilityConverter.
 ### Supported .Net Versions
 - .Net 6
 - .Net 7